Flatten mta-next-train into single-file runtime
